Move open_maybe_zipped to io instead of utils
